darwin-m1: auto generate

+++ ccgo_darwin_arm64.go
@@ -61951,7 +61951,7 @@ func XTclNRTryObjCmd(tls *libc.TLS, clientData TClientData, interp uintptr, objc
 		(*TTcl_Obj)(unsafe.Pointer((*(*[5]uintptr)(unsafe.Pointer(bp + 16)))[int32(1)])).Fbytes = libc.UintptrFromInt32(0)
 		(*TTcl_Obj)(unsafe.Pointer((*(*[5]uintptr)(unsafe.Pointer(bp + 16)))[int32(1)])).FinternalRep.FlongValue = int64(*(*int32)(unsafe.Pointer(bp + 4)))
 		(*TTcl_Obj)(unsafe.Pointer((*(*[5]uintptr)(unsafe.Pointer(bp + 16)))[int32(1)])).FtypePtr = uintptr(unsafe.Pointer(&XtclIntType)) /* returnCode */
-		if (*(*[5]uintptr)(unsafe.Pointer(bp + 16)))[int32(2)] == libc.UintptrFromInt32(0) {                                               /* errorCodePrefix */
+		if (*(*[5]uintptr)(unsafe.Pointer(bp + 16)))[int32(2)] == libc.UintptrFromInt32(0) {                                              /* errorCodePrefix */
 			if XtclFreeObjList == libc.UintptrFromInt32(0) {
 				XTclAllocateFreeObjects(tls)
 			}
@@ -282616,7 +282616,7 @@ func _FileSeekProc(tls *libc.TLS, instanceData TClientData, offset int64, mode i
 	/*
 	 * Save our current place in case we need to roll-back the seek.
 	 */
-	oldLoc = int64(libc.Xlseek(tls, (*TFileState)(unsafe.Pointer(fsPtr)).Ffd, libc.Int64FromInt32(0), int32(m_SEEK_CUR)))
+	oldLoc = libc.Xlseek(tls, (*TFileState)(unsafe.Pointer(fsPtr)).Ffd, libc.Int64FromInt32(0), int32(m_SEEK_CUR))
 	if oldLoc == int64(-libc.Int32FromInt32(1)) {
 		/*
 		 * Bad things are happening. Error out...
@@ -282624,7 +282624,7 @@ func _FileSeekProc(tls *libc.TLS, instanceData TClientData, offset int64, mode i
 		*(*int32)(unsafe.Pointer(errorCodePtr)) = *(*int32)(unsafe.Pointer(libc.X__error(tls)))
 		return -int32(1)
 	}
-	newLoc = int64(libc.Xlseek(tls, (*TFileState)(unsafe.Pointer(fsPtr)).Ffd, offset, mode))
+	newLoc = libc.Xlseek(tls, (*TFileState)(unsafe.Pointer(fsPtr)).Ffd, offset, mode)
 	/*
 	 * Check for expressability in our return type, and roll-back otherwise.
 	 */
@@ -282670,7 +282670,7 @@ func _FileWideSeekProc(tls *libc.TLS, instanceData TClientData, offset TTcl_Wide
 	var v1 int32
 	_, _, _ = fsPtr, newLoc, v1
 	fsPtr = instanceData
-	newLoc = int64(libc.Xlseek(tls, (*TFileState)(unsafe.Pointer(fsPtr)).Ffd, offset, mode))
+	newLoc = libc.Xlseek(tls, (*TFileState)(unsafe.Pointer(fsPtr)).Ffd, offset, mode)
 	if newLoc == int64(-int32(1)) {
 		v1 = *(*int32)(unsafe.Pointer(libc.X__error(tls)))
 	} else {
